@@ -1,11 +1,11 @@
 /* eslint-env node */
-const express = require("express");
-const http = require("http");
-const path = require("path");
+import express from "express";
+import * as http from "http";
+import * as path from "path";
 
 const app = express();
 const server = http.createServer(app);
-const root = path.resolve(__dirname, "../dist/storybook");
+const root = path.resolve(import.meta.dirname, "../dist/storybook");
 app.use(express.static(root));
 const port = process.env.PORT || 9002;
 
