@@ -1,6 +1,6 @@
 /* eslint-env node */
-const fs = require("fs");
-const path = require("path");
+import * as fs from "fs";
+import * as path from "path";
 
 const fileRegex = /\.(vert|frag)$/;
 
