@@ -47,7 +47,7 @@ const config = {
   input: path.join(rootDir, "src", "server", isProduction ? "prod.ts" : "dev.ts"),
   output: {
     dir: path.join(rootDir, "dist"),
-    format: "cjs",
+    format: "es",
     name: "NUsightServer",
     sourcemap: isContinuousIntegration ? false : true,
   },
@@ -59,7 +59,6 @@ const config = {
       limit: 0, // Don't inline any assets, always copy and link
     }),
     typescript({
-      module: "esnext",
       include: ["./src/**/*"],
       exclude: ["./node_modules/**/*", "./src/client/**/*", "./src/tests/**/*"],
       ...(isContinuousIntegration ? { sourceMap: false } : {}),
